backend/pyproject.toml
[project]
name = "manaforge-backend"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "fastapi>=0.111",
    "uvicorn[standard]>=0.30",
    "httpx>=0.27",
    "openai>=1.30",
    "pillow>=10.3",
    "python-multipart>=0.0.9",
    "pydantic-settings>=2.3",
    "sse-starlette>=2.1",
]

